move styles into file
