--- src/components/UI/Modals/UpsertModal/upsertModalStyles.ts
@@ -1,30 +1,0 @@
-import { Style } from "shared/theme";
-
-export const upsertModalStyles: Style = {
-  modal: {
-    position: "absolute",
-    top: "45%",
-    left: "50%",
-    transform: "translate(-50%, -50%)",
-    bgcolor: "white",
-    width: "350px",
-    height: "500px",
-    boxShadow: 24,
-    outline: "none",
-  },
-  innerContainer: {
-    textAlign: "center",
-    display: "flex",
-    flexDirection: "column",
-  },
-  heading: {
-    backgroundColor: "#191a19",
-    py: 2,
-    mb: 2,
-  },
-  icon: {
-    position: "absolute",
-    right: 4,
-    top: 5,
-  },
-};
